requirements.txt
pandas
tqdm
pyyaml
numpy
sentence-transformers
faiss-cpu
scikit-learn
pytest